+++ src/Reservation.Application/Categories/Queries/GetCategory/GetCategoryQueryResponse.cs
@@ -3,5 +3,5 @@ namespace Reservation.Application.Categories.Queries.GetCategory;
 public record GetCategoryQueryResponse
 (
     Guid Id, string Title, string Description,
-    string CoverImagePath, int AveragePoint
+    string CoverImagePath, double AveragePoint
 ) : IResponse;
