+++ src/Reservation.Application/Categories/Queries/GetSubCategoryByCategoryId/GetSubCategoryByCategoryIdQueryResponse.cs
@@ -5,5 +5,5 @@ public record GetSubCategoriesByCategoryIdQueryResponse
     Guid Id, string Title,
     string Description,
     string CoverImagePath,
-    int AveragePoint
+    double AveragePoint
 ) : IResponse;
